rename file, and add check for global assignments
